검색 결과 dispose 원인 파악중
