+++ lib/controller/search_controller.dart
@@ -1,6 +1,9 @@
+import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:village/main.dart';
-import 'package:village/view/pages/search/result_page/search_result_page_view_model.dart';
+import 'package:village/model/search/search.dart';
+import 'package:village/view/pages/search/result_page/search_result_page.dart';
+import 'package:village/view/pages/search/search_page_view_model.dart';
 
 final searchControllerProvider = Provider<SearchController>((ref) {
   return SearchController(ref);
@@ -11,7 +14,14 @@ class SearchController {
   final Ref ref;
   SearchController(this.ref);
 
-  Future<void> refresh(String keyword) async{
-    ref.read(searchResultPageProvider.notifier).notifyInit(keyword);
+  Future<void> searchKeyword(String keyword) async {
+    Navigator.push(
+      mContext!,
+      MaterialPageRoute(
+          builder: (context) => SearchResultPage(
+                keyword: keyword,
+              )),
+    );
+    ref.read(searchPageProvider.notifier).addSearch(Search(keyword: keyword));
   }
 }
