+++ lib/view/pages/search/keyword_page/search_keyword_page.dart
@@ -1,19 +1,17 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:village/controller/search_controller.dart';
 import 'package:village/view/pages/search/keyword_page/widgets/search_keyword_bar.dart';
 import 'package:village/view/pages/search/keyword_page/widgets/search_keyword_body.dart';
 
-class SearchKeywordPage extends StatefulWidget {
-  const SearchKeywordPage({super.key});
+class SearchKeywordPage extends ConsumerWidget {
+  SearchKeywordPage({super.key});
 
-  @override
-  State<SearchKeywordPage> createState() => _SearchKeywordPageState();
-}
-
-class _SearchKeywordPageState extends State<SearchKeywordPage> {
   final TextEditingController _searchController = TextEditingController();
 
   @override
-  Widget build(BuildContext context) {
+  Widget build(BuildContext context, WidgetRef ref) {
+    final cp = ref.read(searchControllerProvider);
     return Scaffold(
       appBar: AppBar(
         elevation: 0,
@@ -21,6 +19,7 @@ class _SearchKeywordPageState extends State<SearchKeywordPage> {
         title: Row(
           children: [
             SearchKeywordBar(
+                cp: cp,
                 searchController: _searchController,
                 text: '지역, 공간 카테고리로 찾아보세요'),
             const Spacer(),
@@ -35,7 +34,7 @@ class _SearchKeywordPageState extends State<SearchKeywordPage> {
           ],
         ),
       ),
-      body: SearchKeywordBody(),
+      body: const SearchKeywordBody(),
     );
   }
 }
