+++ lib/view/pages/search/keyword_page/widgets/search_keyword_bar.dart
@@ -1,19 +1,18 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:logger/logger.dart';
 import 'package:village/core/constants/size.dart';
-import 'package:village/model/search/search.dart';
-import 'package:village/view/pages/search/keyword_page/search_keyword_page_view_model.dart';
-import 'package:village/view/pages/search/result_page/search_result_page.dart';
 
-class SearchKeywordBar extends ConsumerWidget{
-  const SearchKeywordBar( {
+class SearchKeywordBar extends ConsumerWidget {
+  const SearchKeywordBar({
     required this.text,
     super.key,
     required TextEditingController searchController,
+    required this.cp,
   }) : _searchController = searchController;
+
   final text;
   final TextEditingController _searchController;
+  final cp;
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
@@ -25,15 +24,9 @@ class SearchKeywordBar extends ConsumerWidget{
           color: Colors.grey[200], borderRadius: BorderRadius.circular(24)),
       child: TextField(
         controller: _searchController,
-        onSubmitted: (String keyword) async{
+        onSubmitted: (String keyword) async {
           if (keyword.isNotEmpty) {
-            final viewModel = ref.read(searchKeywordPageProvider.notifier);
-            await viewModel.addSearch(Search(keyword: keyword));
-            Logger().d(keyword);
-            Navigator.push(
-              context,
-              MaterialPageRoute(builder: (context) => SearchResultPage(keyword: keyword)),
-            );
+            cp.searchKeyword(keyword);
           }
         },
         decoration: InputDecoration(
@@ -44,5 +37,4 @@ class SearchKeywordBar extends ConsumerWidget{
       ),
     );
   }
-
 }
