+++ lib/view/pages/search/result_page/search_result_page.dart
@@ -1,29 +1,29 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:village/model/search/search.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:village/controller/search_controller.dart';
 import 'package:village/view/pages/search/keyword_page/widgets/search_keyword_bar.dart';
+import 'package:village/view/pages/search/result_page/search_result_page_view_model.dart';
 import 'package:village/view/pages/search/result_page/widgets/search_result_body.dart';
 import 'package:village/view/widgets/custom_back_icon.dart';
 
-class SearchResultPage extends StatefulWidget {
+class SearchResultPage extends ConsumerWidget {
+  String keyword;
 
-  final String keyword;
+  SearchResultPage({super.key, this.keyword = ''});
 
-  const SearchResultPage({super.key, required this.keyword});
-
-  @override
-  State<SearchResultPage> createState() => _SearchResultPageState();
-}
-
-class _SearchResultPageState extends State<SearchResultPage> {
   final TextEditingController _searchController = TextEditingController();
   @override
-  Widget build(BuildContext context) {
+  Widget build(BuildContext context, WidgetRef ref) {
+    final cp = ref.read(searchControllerProvider);
+    ref.read(searchResultPageProvider.notifier).notifyInit(keyword);
+    final pm = ref.watch(searchResultPageProvider);
     return Scaffold(
       appBar: AppBar(
         leading: CustomBackIcon(context),
         actions: [
-          SearchKeywordBar(searchController: _searchController, text: widget.keyword),
+          SearchKeywordBar(
+              cp: cp, searchController: _searchController, text: keyword),
           const SizedBox(
             width: 20,
           ),
@@ -34,7 +34,7 @@ class _SearchResultPageState extends State<SearchResultPage> {
         backgroundColor: Colors.transparent,
         elevation: 0.0,
       ),
-      body: const SearchResultBody(),
+      body: SearchResultBody(pm: pm),
     );
   }
 }
